Create an Order Feature (#10)
* Add required elements to navbar.html

* Create Order first version.

* Thymeleaf displays create-order.html correctly.

* Display and fetch quantity correctly.

* Initial post method for creating an order.

* POST method works correctly.

+++ backend/src/main/java/mk/ukim/finki/db/distributorapp/model/dto/CategoryDto.java
@@ -2,11 +2,10 @@ package mk.ukim.finki.db.distributorapp.model.dto;
 
 import lombok.AllArgsConstructor;
 import lombok.Data;
-import lombok.NoArgsConstructor;
 
 @Data
 @AllArgsConstructor
 public class CategoryDto {
-    private Long id;
+    private Integer id;
     private String name;
 }
